Render semantic version

Render a semantic version list or a semantic version pointer as an R list giving the major, minor and patch version as an integer and the prerelease and build version as a charcter

render_version(x, ...)

# S3 method for svptr
render_version(x, ...)

# S3 method for svlist
render_version(x, ...)

Arguments

x

An object of class "svlist" or "svptr".

...

Additional arguments passed on to methods.

Value

A list of semantic versions represented as integers (major, minor, patch) and characters (prerelease, build)

Examples

semver <- parse_version(c("1.2.3", "1.2.4", "1.0.0", "1.2.4-beta+2", "2.1.3-alpha", "1.9.4"))
render_version(semver[[4]])

#> $major
#> [1] 1
#> 
#> $minor
#> [1] 2
#> 
#> $patch
#> [1] 4
#> 
#> $prerelease
#> [1] "beta"
#> 
#> $build
#> [1] "2"
#> 

render_version(semver)

#> [[1]]
#> [[1]]$major
#> [1] 1
#> 
#> [[1]]$minor
#> [1] 2
#> 
#> [[1]]$patch
#> [1] 3
#> 
#> [[1]]$prerelease
#> [1] ""
#> 
#> [[1]]$build
#> [1] ""
#> 
#> 
#> [[2]]
#> [[2]]$major
#> [1] 1
#> 
#> [[2]]$minor
#> [1] 2
#> 
#> [[2]]$patch
#> [1] 4
#> 
#> [[2]]$prerelease
#> [1] ""
#> 
#> [[2]]$build
#> [1] ""
#> 
#> 
#> [[3]]
#> [[3]]$major
#> [1] 1
#> 
#> [[3]]$minor
#> [1] 0
#> 
#> [[3]]$patch
#> [1] 0
#> 
#> [[3]]$prerelease
#> [1] ""
#> 
#> [[3]]$build
#> [1] ""
#> 
#> 
#> [[4]]
#> [[4]]$major
#> [1] 1
#> 
#> [[4]]$minor
#> [1] 2
#> 
#> [[4]]$patch
#> [1] 4
#> 
#> [[4]]$prerelease
#> [1] "beta"
#> 
#> [[4]]$build
#> [1] "2"
#> 
#> 
#> [[5]]
#> [[5]]$major
#> [1] 2
#> 
#> [[5]]$minor
#> [1] 1
#> 
#> [[5]]$patch
#> [1] 3
#> 
#> [[5]]$prerelease
#> [1] "alpha"
#> 
#> [[5]]$build
#> [1] ""
#> 
#> 
#> [[6]]
#> [[6]]$major
#> [1] 1
#> 
#> [[6]]$minor
#> [1] 9
#> 
#> [[6]]$patch
#> [1] 4
#> 
#> [[6]]$prerelease
#> [1] ""
#> 
#> [[6]]$build
#> [1] ""
#> 
#>